Notification commits (CRUD commit)

+++ src/main/java/com/example/insuranceprototype/Notification/NotificationRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface NotificationRepository  extends JpaRepository<Notification, Long> {
+public interface NotificationRepository  extends JpaRepository<Notification, Integer> {
 
     List<Notification> findByEmployeeId(Long employeeId);
 }
